+++ Over-and-Under/include/odometry.h
@@ -3,12 +3,12 @@
 namespace eftl{
 class Odometry{
     public:
-        float encoderTravel;
-        float encoderDeg; 
-        float encoderDelta;
-        float prevEncoder;
-Odometry(float encoder);
+        long double encoderTravel;
+        long double encoderDeg; 
+        long double encoderDelta;
+        long double prevEncoder;
+Odometry(long double  encoder);
 static std::array<double,2> tracking();
-static int trackingCallback();
+static  int trackingCallback();
 };
 }
